Move _create_flaky_qstat to where it is used
